requirements.txt
streamlit==1.39.0
plotly==5.24.1
pandas==2.2.3
python-dotenv==1.0.1
groq==0.13.0
langchain==0.3.7
langchain-groq==0.2.1
chromadb==0.5.23
sentence-transformers==3.3.1
